+++ sentry-ruby/spec/sentry/scope_spec.rb
@@ -45,7 +45,7 @@ RSpec.describe Sentry::Scope do
       copy.set_transaction_name("foo", source: :url)
       copy.fingerprint << "bar"
 
-      expect(subject.breadcrumbs.to_hash).to eq({ values: [] })
+      expect(subject.breadcrumbs.to_h).to eq({ values: [] })
       expect(subject.contexts[:os].keys).to match_array([:name, :version, :build, :kernel_version, :machine])
       expect(subject.contexts.dig(:runtime, :version)).to match(/ruby/)
       expect(subject.extra).to eq({})
@@ -246,7 +246,7 @@ RSpec.describe Sentry::Scope do
 
         scope.apply_to_event(log_event)
 
-        log_hash = log_event.to_hash
+        log_hash = log_event.to_h
 
         expect(log_hash[:attributes]["user.id"]).to eq(123)
         expect(log_hash[:attributes]["user.name"]).to eq("john_doe")
@@ -257,7 +257,7 @@ RSpec.describe Sentry::Scope do
         scope = described_class.new
         scope.apply_to_event(log_event)
 
-        log_hash = log_event.to_hash
+        log_hash = log_event.to_h
 
         expect(log_hash[:attributes]).not_to have_key("user.id")
         expect(log_hash[:attributes]).not_to have_key("user.name")
@@ -351,7 +351,7 @@ RSpec.describe Sentry::Scope do
       it "sets the request info the Event" do
         subject.apply_to_event(event)
 
-        expect(event.to_hash.dig(:request, :url)).to eq("http://example.org/test")
+        expect(event.to_h.dig(:request, :url)).to eq("http://example.org/test")
       end
     end
   end