+++ sentry-ruby/spec/sentry/transaction_spec.rb
@@ -392,9 +392,9 @@ RSpec.describe Sentry::Transaction do
     end
   end
 
-  describe "#to_hash" do
+  describe "#to_h" do
     it "returns correct data" do
-      hash = subject.to_hash
+      hash = subject.to_h
 
       expect(hash[:op]).to eq("sql.query")
       expect(hash[:description]).to eq("SELECT * FROM users;")
@@ -421,7 +421,7 @@ RSpec.describe Sentry::Transaction do
       subject.finish
 
       expect(events.count).to eq(1)
-      event = events.last.to_hash
+      event = events.last.to_h
 
       # don't contain itself
       expect(event[:spans]).to be_empty
@@ -431,7 +431,7 @@ RSpec.describe Sentry::Transaction do
       subject.finish
 
       expect(events.count).to eq(1)
-      event = events.last.to_hash
+      event = events.last.to_h
 
       expect(event[:transaction]).to eq("foo")
     end
@@ -441,7 +441,7 @@ RSpec.describe Sentry::Transaction do
       subject.finish(end_timestamp: timestamp)
 
       expect(events.count).to eq(1)
-      event = events.last.to_hash
+      event = events.last.to_h
 
       expect(event[:timestamp]).to eq(timestamp)
     end
@@ -454,7 +454,7 @@ RSpec.describe Sentry::Transaction do
       subject.finish
 
       expect(events.count).to eq(1)
-      event = events.last.to_hash
+      event = events.last.to_h
 
       expect(event[:tags]).to eq({ foo: 'bar', name: "apple" })
     end
@@ -550,7 +550,7 @@ RSpec.describe Sentry::Transaction do
         subject.set_measurement("metric.foo", 0.5, "second")
         subject.finish
 
-        transaction = events.last.to_hash
+        transaction = events.last.to_h
         expect(transaction[:measurements]).to eq(
           { "metric.foo" => { value: 0.5, unit: "second" } }
         )